+++ src/routes/ConsortiumManager/utils/index.js
@@ -1,2 +1,3 @@
 export { groupMembersSettings } from './groupMembersSettings';
 export { validateNameAndCodeRequired } from './validateNameAndCodeRequired';
+export { validateNameRequired } from './validateNameRequired';