+++ src/routes/ConsortiumManager/utils/validateNameRequired.js
@@ -1,0 +1,3 @@
+import { validateRequired } from '../../../components/ConsortiaControlledVocabulary/validators';
+
+export const validateNameRequired = (item) => ({ name: validateRequired({ value: item.name }) });